refactor(mcp): split workspace mod.rs into smaller modules
